review: split accord_space_settings.dart, add coverage for new pure logic
PR review findings, addressed:

- accord_space_settings.dart grew to 1231 lines during this same PR's
  build() decomposition (the section widgets stayed in one file). Split
  into a part-file library matching the accord_home.dart/message_pane/
  pattern already used elsewhere in this branch: the entry point + state
  (573 lines), form sections (438), management sections (154), and the
  adaptive layout (73). Pure move, verified byte-identical modulo the new
  part directives.
- Added unit tests for three previously-untested pure-logic extractions
  introduced by the audit: channel_reorder.dart (flatten/diff behind both
  drag-reorder surfaces), self_loading_list.dart (the two pagination
  mixins now backing 5 moderation screens), and participant_display.dart
  (identity resolution shared by the voice lobby, grid, and text panel).
- Documented the space-settings split as a G1 follow-up in
  docs/complexity-audit.md's implementation record.

The voice.dart reconnect-gating change flagged by review turned out to be
an already-documented, already-tested intentional fix (see the B1 [Δ edge
case] entry in complexity-audit.md), not a regression — no change needed.
